moved utils out of conn manager
